Split sgemm for lora_a and lora_b
