+++ python/sglang/srt/lora/backend/__init__.py
@@ -1,3 +1,4 @@
+from .base_backend import BaseLoraBackend
 from .flashinfer_backend import FlashInferLoraBackend
 from .triton_backend import TritonLoraBackend
