+++ python/sglang/srt/lora/backend/flashinfer_backend.py
@@ -1,7 +1,7 @@
 import torch
 from flashinfer import SegmentGEMMWrapper
 
-from sglang.srt.lora.backend.base_backend import BaseLoraBackend
+from sglang.srt.lora.backend import BaseLoraBackend
 from sglang.srt.lora.lora import LoraBatchInfo
 
 
@@ -15,7 +15,18 @@ class FlashInferLoraBackend(BaseLoraBackend):
         workspace_buffer = torch.empty(1 * 1024 * 1024, dtype=torch.int8, device="cuda")
         self.segment_gemm = SegmentGEMMWrapper(workspace_buffer)
 
-    def run_sgemm(self, x: torch.Tensor, weights: torch.Tensor) -> torch.Tensor:
+    def run_lora_a_sgemm(self, x: torch.Tensor, weights: torch.Tensor) -> torch.Tensor:
+
+        return self.segment_gemm.run(
+            x=x,
+            weights=weights,
+            batch_size=self.batch_info.bs,
+            weight_column_major=True,
+            seg_indptr=self.batch_info.seg_indptr,
+            weight_indices=self.batch_info.weight_indices,
+        )
+
+    def run_lora_b_sgemm(self, x: torch.Tensor, weights: torch.Tensor) -> torch.Tensor:
 
         return self.segment_gemm.run(
             x=x,
@@ -35,7 +46,7 @@ class FlashInferLoraBackend(BaseLoraBackend):
     ) -> torch.Tensor:
 
         # Shape of lora_a_output: (s, 3 * r)
-        lora_a_output = self.run_sgemm(x=x, weights=qkv_lora_a)
+        lora_a_output = self.run_lora_a_sgemm(x=x, weights=qkv_lora_a)
 
         lora_rank = kv_lora_b.shape[-1]
         output_dim_q = q_lora_b.shape[-2]
@@ -48,19 +59,19 @@ class FlashInferLoraBackend(BaseLoraBackend):
 
         # FIXME parallelize qkv
         # q
-        lora_output[:, :output_dim_q] = self.run_sgemm(
+        lora_output[:, :output_dim_q] = self.run_lora_b_sgemm(
             x=lora_a_output[:, :lora_rank].contiguous(), weights=q_lora_b[0]
         )
 
         # kv
-        lora_output[:, output_dim_q : output_dim_q + output_dim_kv] = self.run_sgemm(
+        lora_output[:, output_dim_q : output_dim_q + output_dim_kv] = self.run_lora_b_sgemm(
             x=lora_a_output[:, lora_rank : 2 * lora_rank].contiguous(),
             weights=kv_lora_b[0],
         )
 
         lora_output[
             :, output_dim_q + output_dim_kv : output_dim_q + 2 * output_dim_kv
-        ] = self.run_sgemm(
+        ] = self.run_lora_b_sgemm(
             x=lora_a_output[:, 2 * lora_rank : 3 * lora_rank].contiguous(),
             weights=kv_lora_b[1],
         )
